requirements.txt
faiss-cpu==1.8.0
fastapi==0.111.0
mistralai==0.4.0
numpy==1.26.4
PyPDF2==3.0.1
python-dotenv==1.0.1
pydantic==2.7.1
uvicorn==0.29.0
